added different topbars and bottom bars
