+++ app/src/main/java/com/vivich/starlitreadingapp/ui/screens/lobby/bookContent/BottomNavBar.kt
@@ -1,4 +1,4 @@
-package com.vivich.starlitreadingapp.ui.screens.classic.bookContent
+package com.vivich.starlitreadingapp.ui.screens.lobby.bookContent
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
