+++ app/src/main/java/com/vivich/starlitreadingapp/ui/screens/lobby/bookDetails/BookDetailsScreen.kt
@@ -1,0 +1,19 @@
+package com.vivich.starlitreadingapp.ui.screens.lobby.bookDetails
+
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+import com.vivich.starlitreadingapp.BookDetailsScreen
+
+@Composable
+fun BookDetailsScreen(
+    navHostController: NavHostController,
+){
+    Scaffold(
+        topBar = {},
+        bottomBar = {},
+        content = {
+
+        }
+    )
+}
