+++ app/src/main/java/com/vivich/starlitreadingapp/ui/screens/lobby/bookDetails/TopNavBar.kt
@@ -1,4 +1,4 @@
-package com.vivich.starlitreadingapp.ui.screens.classic.bookDetails
+package com.vivich.starlitreadingapp.ui.screens.lobby.bookDetails
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
